backend/requirements.txt
Flask==3.1.3
Flask-Cors==6.0.2
yfinance==1.2.0
vaderSentiment==3.3.2
requests==2.33.0
python-dotenv==1.2.1
praw==7.8.1
psycopg2==2.9.11